Port ItemCommand
